Migrate features/content/searchButton to TypeScript
- drop nonsensical await before function reference in addEventListener
- wrap async click handler with void to satisfy listener signature
- getLinkMeta uses single courses/X/lessons/Y match with null guard
- inline tiny helpers (getConspectData, getConspectTitle, icon element)
