+++ src/chrome-scripts/features/content/index.js
@@ -2,5 +2,5 @@ export { createHomeworkObserver as homeworkPercent } from './homeworkPercent';
 export { createWebinarObserver as webinarPercent } from './webinarPercent';
 export { createReadingTimeObserver as readingTime } from './readingTime';
 export { default as readingList } from './readingList.js';
-export { default as searchButton } from './searchButton.js';
+export { createSearchButtonObserver as searchButton } from './searchButton';
 export { fixYellowBlocks } from './fixYellowBlocks';
